+++ src/app/pages/user/bucket/bucket-datatable-page/bucket-datatable-page.component.ts
@@ -129,6 +129,13 @@ export class BucketDatatablePageComponent {
         }
       },
       {
+        title: TEXT('Lifecycle'),
+        icon: 'mdi mdi-clock-end',
+        callback: () => {
+          this.router.navigate([`/buckets/lifecycle/${bucket.Name}`]);
+        }
+      },
+      {
         type: 'divider'
       },
       {