+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add button to copy the current path of the object browser to the clipboard.
 - Support legal holds on objects (gh#aquarist-labs/s3gw#365).
+- Add support to manage lifecycle rules per bucket (gh#aquarist-labs/s3gw#349).
 
 ### Changed